update test runner to cast longs to i32

+++ integration_test/src/test_runner/mod.rs
@@ -709,7 +709,7 @@ fn get_data(
                                 [0..*out_len_or_ind as usize]
                                 .to_string());
                         }
-                        CDataType::SLong => data = json!(*(buffer as *const i64)),
+                        CDataType::SLong => data = json!(*(buffer as *const i32)),
                         CDataType::Double => data = json!(*(buffer as *const f64)),
                         CDataType::Bit => data = json!(*(buffer as *const bool)),
                         _ => {}
